+++ cli/npm/managed/installers/common/bin_entries.rs
