+++ cli/npm/managed/installers/common/lifecycle_scripts.rs
