+++ cli/npm/managed/mod.rs
@@ -28,11 +28,14 @@ use deno_runtime::colors;
 use deno_runtime::ops::process::NpmProcessStateProvider;
 use deno_semver::package::PackageNv;
 use deno_semver::package::PackageReq;
+use installer::AddPkgReqsResult;
+use installer::NpmResolutionInstaller;
+use installers::create_npm_fs_installer;
+use installers::NpmPackageFsInstaller;
 use node_resolver::errors::PackageFolderResolveError;
 use node_resolver::errors::PackageFolderResolveIoError;
 use node_resolver::InNpmPackageChecker;
 use node_resolver::NpmPackageFolderResolver;
-use resolution::AddPkgReqsResult;
 
 use self::resolution::NpmResolution;
 use self::resolvers::create_npm_fs_resolver;
@@ -55,6 +58,8 @@ use crate::sys::CliSys;
 use crate::util::progress_bar::ProgressBar;
 use crate::util::sync::AtomicFlag;
 
+mod installer;
+mod installers;
 mod resolution;
 mod resolvers;
 
@@ -156,11 +161,7 @@ fn create_inner(
   snapshot: Option<ValidSerializedNpmResolutionSnapshot>,
   lifecycle_scripts: LifecycleScriptsConfig,
 ) -> Arc<dyn CliNpmResolver> {
-  let resolution = Arc::new(NpmResolution::from_serialized(
-    registry_info_provider.clone(),
-    snapshot,
-    maybe_lockfile.clone(),
-  ));
+  let resolution = Arc::new(NpmResolution::from_serialized(snapshot));
   let tarball_cache = Arc::new(CliNpmTarballCache::new(
     npm_cache.clone(),
     http_client,
@@ -168,18 +169,25 @@ fn create_inner(
     npm_rc.clone(),
   ));
 
-  let fs_resolver = create_npm_fs_resolver(
+  let fs_installer = create_npm_fs_installer(
     npm_cache.clone(),
     &npm_install_deps_provider,
     &text_only_progress_bar,
     resolution.clone(),
     sys.clone(),
     tarball_cache.clone(),
-    node_modules_dir_path,
+    node_modules_dir_path.clone(),
     npm_system_info.clone(),
     lifecycle_scripts.clone(),
   );
+  let fs_resolver = create_npm_fs_resolver(
+    npm_cache.clone(),
+    resolution.clone(),
+    sys.clone(),
+    node_modules_dir_path,
+  );
   Arc::new(ManagedCliNpmResolver::new(
+    fs_installer,
     fs_resolver,
     maybe_lockfile,
     registry_info_provider,
@@ -301,6 +309,7 @@ pub enum PackageCaching<'a> {
 /// An npm resolver where the resolution is managed by Deno rather than
 /// the user bringing their own node_modules (BYONM) on the file system.
 pub struct ManagedCliNpmResolver {
+  fs_installer: Arc<dyn NpmPackageFsInstaller>,
   fs_resolver: Arc<dyn NpmPackageFsResolver>,
   maybe_lockfile: Option<Arc<CliLockfile>>,
   registry_info_provider: Arc<CliNpmRegistryInfoProvider>,
@@ -308,6 +317,7 @@ pub struct ManagedCliNpmResolver {
   npm_install_deps_provider: Arc<NpmInstallDepsProvider>,
   sys: CliSys,
   resolution: Arc<NpmResolution>,
+  resolution_installer: NpmResolutionInstaller,
   tarball_cache: Arc<CliNpmTarballCache>,
   text_only_progress_bar: ProgressBar,
   npm_system_info: NpmSystemInfo,
@@ -348,6 +358,7 @@ pub enum ResolvePkgFolderFromDenoModuleError {
 impl ManagedCliNpmResolver {
   #[allow(clippy::too_many_arguments)]
   pub fn new(
+    fs_installer: Arc<dyn NpmPackageFsInstaller>,
     fs_resolver: Arc<dyn NpmPackageFsResolver>,
     maybe_lockfile: Option<Arc<CliLockfile>>,
     registry_info_provider: Arc<CliNpmRegistryInfoProvider>,
@@ -360,7 +371,13 @@ impl ManagedCliNpmResolver {
     npm_system_info: NpmSystemInfo,
     lifecycle_scripts: LifecycleScriptsConfig,
   ) -> Self {
+    let resolution_installer = NpmResolutionInstaller::new(
+      registry_info_provider.clone(),
+      resolution.clone(),
+      maybe_lockfile.clone(),
+    );
     Self {
+      fs_installer,
       fs_resolver,
       maybe_lockfile,
       registry_info_provider,
@@ -368,6 +385,7 @@ impl ManagedCliNpmResolver {
       npm_install_deps_provider,
       text_only_progress_bar,
       resolution,
+      resolution_installer,
       sys,
       tarball_cache,
       npm_system_info,
@@ -489,7 +507,7 @@ impl ManagedCliNpmResolver {
       };
     }
 
-    let mut result = self.resolution.add_package_reqs(packages).await;
+    let mut result = self.resolution_installer.add_package_reqs(packages).await;
 
     if result.dependencies_result.is_ok() {
       if let Some(lockfile) = self.maybe_lockfile.as_ref() {
@@ -512,7 +530,7 @@ impl ManagedCliNpmResolver {
     &self,
     packages: &[PackageReq],
   ) -> Result<(), AnyError> {
-    self.resolution.set_package_reqs(packages).await
+    self.resolution_installer.set_package_reqs(packages).await
   }
 
   pub fn snapshot(&self) -> NpmResolutionSnapshot {
@@ -554,7 +572,7 @@ impl ManagedCliNpmResolver {
     &self,
     caching: PackageCaching<'_>,
   ) -> Result<(), JsErrorBox> {
-    self.fs_resolver.cache_packages(caching).await
+    self.fs_installer.cache_packages(caching).await
   }
 
   pub fn resolve_pkg_folder_from_deno_module(
@@ -738,14 +756,11 @@ impl CliNpmResolver for ManagedCliNpmResolver {
 
   fn clone_snapshotted(&self) -> Arc<dyn CliNpmResolver> {
     // create a new snapshotted npm resolution and resolver
-    let npm_resolution = Arc::new(NpmResolution::new(
-      self.registry_info_provider.clone(),
-      self.resolution.snapshot(),
-      self.maybe_lockfile.clone(),
-    ));
+    let npm_resolution =
+      Arc::new(NpmResolution::new(self.resolution.snapshot()));
 
     Arc::new(ManagedCliNpmResolver::new(
-      create_npm_fs_resolver(
+      create_npm_fs_installer(
         self.npm_cache.clone(),
         &self.npm_install_deps_provider,
         &self.text_only_progress_bar,
@@ -755,6 +770,12 @@ impl CliNpmResolver for ManagedCliNpmResolver {
         self.root_node_modules_path().map(ToOwned::to_owned),
         self.npm_system_info.clone(),
         self.lifecycle_scripts.clone(),
+      ),
+      create_npm_fs_resolver(
+        self.npm_cache.clone(),
+        npm_resolution.clone(),
+        self.sys.clone(),
+        self.root_node_modules_path().map(ToOwned::to_owned),
       ),
       self.maybe_lockfile.clone(),
       self.registry_info_provider.clone(),
