+++ cli/util/sync/mod.rs
@@ -1,11 +1,9 @@
 // Copyright 2018-2025 the Deno authors. MIT license.
 
 mod async_flag;
-mod sync_read_async_write_lock;
 mod task_queue;
 
 pub use async_flag::AsyncFlag;
 pub use deno_core::unsync::sync::AtomicFlag;
-pub use sync_read_async_write_lock::SyncReadAsyncWriteLock;
 pub use task_queue::TaskQueue;
 pub use task_queue::TaskQueuePermit;
